Prevent users from modifying clearances above theirs

+++ src/main/kotlin/me/mrkirby153/KirBot/command/executors/clearance/CommandOverrideClearance.kt
@@ -1,11 +1,13 @@
 package me.mrkirby153.KirBot.command.executors.clearance
 
 import me.mrkirby153.KirBot.command.CommandException
+import me.mrkirby153.KirBot.command.CommandManager
 import me.mrkirby153.KirBot.command.args.CommandContext
 import me.mrkirby153.KirBot.command.executors.CmdExecutor
 import me.mrkirby153.KirBot.database.api.ClearanceOverride
 import me.mrkirby153.KirBot.user.Clearance
 import me.mrkirby153.KirBot.utils.Context
+import me.mrkirby153.KirBot.utils.getClearance
 
 class CommandOverrideClearance : CmdExecutor() {
     override fun execute(context: Context, cmdContext: CommandContext) {
@@ -18,6 +20,16 @@ class CommandOverrideClearance : CmdExecutor() {
             it.command.equals(cmd, true)
         }
 
+        val cmdSpec = CommandManager.findCommand(cmd) ?: throw CommandException("That command isn't registered")
+
+        if(context.user.getClearance(context.guild).value < cmdSpec.clearance.value){
+            throw CommandException("You do not have permission to modify this command")
+        }
+
+        val targetClearance = Clearance.valueOf(clearance.toUpperCase())
+        if(!reset && targetClearance.value > context.user.getClearance(context.guild).value){
+            throw CommandException("You cannot change clearance to one higher than yours")
+        }
         if (existingClearance != null) {
             if (reset) {
                 existingClearance.delete().queue {
@@ -29,7 +41,7 @@ class CommandOverrideClearance : CmdExecutor() {
                 return
             } else {
                 try {
-                    existingClearance.clearance = Clearance.valueOf(clearance.toUpperCase())
+                    existingClearance.clearance = targetClearance
                 } catch (e: Exception) {
                     context.send().error("Invalid clearance level").queue()
                 }
@@ -39,7 +51,7 @@ class CommandOverrideClearance : CmdExecutor() {
             }
         } else {
             try {
-                val c = Clearance.valueOf(clearance.toUpperCase())
+                val c = targetClearance
                 ClearanceOverride.create(context.guild, cmd.toLowerCase(), c).queue {
                     context.shard.clearanceOverrides[context.guild.id].add(it)
                     context.send().success("Set clearance to `$clearance`").queue()
